evals/pyproject.toml
[project]
name = "temporal-evals"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "temporalio>=1.11.0",
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["temporal_evals"]

[project.scripts]
eval-worker = "temporal_evals.worker:main_sync"
eval-run = "temporal_evals.run_evals:main_sync"
eval-dashboard = "temporal_evals.dashboard:main_sync"
